style/rust_sodium-sys: move 'init_with_rng' to separate file and avoid using 'include!'
