+++ Cargo.toml
@@ -11,16 +11,16 @@ repository = "https://github.com/maidsafe/rust_sodium"
 version = "0.8.0"
 
 [dependencies]
-libc = "~0.2.38"
+libc = "~0.2.40"
 rand = "~0.4.2"
 rust_sodium-sys = { path = "rust_sodium-sys", version = "~0.8.1" }
-serde = "~1.0.27"
+serde = "~1.0.37"
 unwrap = "~1.1.0"
 
 [dev-dependencies]
 hex = "~0.3.1"
 rmp-serde = "~0.13.7"
-serde_json = "~1.0.10"
+serde_json = "~1.0.13"
 
 [features]
 benchmarks = []
