+++ rust_sodium-sys/Cargo.toml
@@ -13,17 +13,17 @@ version = "0.8.1"
 
 [dependencies]
 lazy_static = "~1.0.0"
-libc = "~0.2.38"
+libc = "~0.2.40"
 rand = "~0.4.2"
 unwrap = "~1.1.0"
 
 [build-dependencies]
-cc = "~1.0.5"
+cc = "~1.0.9"
 flate2 = "~1.0.1"
-libc = "~0.2.38"
+libc = "~0.2.40"
 pkg-config = "~0.3.9"
 reqwest = "~0.8.5"
 sha2 = "~0.7.0"
-tar = "~0.4.14"
+tar = "~0.4.15"
 unwrap = "~1.1.0"
 zip = "~0.3.1"
