+++ rust_sodium-sys/src/bindgen.rs
@@ -1,5 +1,7 @@
 /* automatically generated by rust-bindgen */
 
+use libc;
+
 pub const SODIUM_VERSION_STRING: *const libc::c_char = (b"1.0.16\0" as *const libc::c_uchar) as
     *const libc::c_char;
 pub const SODIUM_LIBRARY_VERSION_MAJOR: libc::c_uint = 10;
